Update to be on par with the cdk main branch.

+++ base/standard/src/main/java/org/openscience/cdk/tools/manipulator/AtomContainerSetManipulator.java
@@ -34,8 +34,6 @@ import org.openscience.cdk.interfaces.IChemObject;
 import org.openscience.cdk.interfaces.IElectronContainer;
 
 /**
- * @cdk.module standard
- * @cdk.githash
  *
  * @see ChemModelManipulator
  */
@@ -112,7 +110,7 @@ public class AtomContainerSetManipulator {
         for (int i = 0; i < set.getAtomContainerCount(); i++) {
             int thisCharge = AtomContainerManipulator.getTotalFormalCharge(set.getAtomContainer(i));
             double stoich = set.getMultiplier(i);
-            charge += stoich * thisCharge;
+            charge += stoich * (double)thisCharge;
         }
         return charge;
     }
@@ -126,7 +124,7 @@ public class AtomContainerSetManipulator {
         for (int i = 0; i < set.getAtomContainerCount(); i++) {
             int thisCharge = AtomContainerManipulator.getTotalFormalCharge(set.getAtomContainer(i));
             double stoich = set.getMultiplier(i);
-            charge += stoich * thisCharge;
+            charge += (int)(stoich * (double)thisCharge);
         }
         return charge;
     }
